2021 - New template for tests - reporting duration instead of millis and support n iterations.  Also - only edit the expected results once - in top of test-file.

+++ src/main/kotlin/no/rodland/advent_2021/Day10.kt
@@ -2,11 +2,11 @@ package no.rodland.advent_2021
 
 @Suppress("UNUSED_PARAMETER")
 object Day10 {
-    fun partOne(list: List<String>): Int {
+    fun partOne(list: List<String>): Long {
         return 2
     }
 
-    fun partTwo(list: List<String>): Int {
+    fun partTwo(list: List<String>): Long {
         return 2
     }
 }
